fix(governance): quorum AG en Decimal et unification Art. 3.87 §5 (#661)
Le quorum d'AG hybride etait calcule et compare en `f64` sur un endpoint
expose (`GET /ag-sessions/{id}/quorum`), et ne verifiait que le volet
quotites — pas le volet « tetes » exige par l'Art. 3.87 §5 CC.

- `AgSession` en `Decimal` de bout en bout (entite, DTO, use case, handler,
  repository). Aucune migration : les colonnes etaient deja `NUMERIC`, le
  repository les degradait par 8 casts `::FLOAT8`, desormais retires.
- Une seule implementation d'Art. 3.87 §5 : la regle du quorum double est
  extraite en predicats partages sur `Meeting` (`double_quorum_reached`),
  appliques par le chemin presentiel (H9) comme par le chemin hybride.
  Changement de comportement : les quotites seules n'emportent plus le
  quorum d'une AG hybride.
- Les deux bornes de l'article sont codifiees distinctement : quotites
  « au moins la moitie » (inclusif), tetes « plus de la moitie » (strict).
- `validate_proxy_mandate` supprimee : morte, en `f64`, et juridiquement
  fausse (cumul en ET de deux regles dont la seconde est une exception).
  La regle reelle est `validate_proxy_limit`, deja en `Decimal` et cablee.
- `BudgetVarianceResponse` / `BudgetStatsResponse` tranches en `Decimal` :
  charges de copropriete, qu'ADR-0008 §A exclut de tout carve-out `f64`.
- Nouveau gate CI `scripts/check-no-f64-money.sh` (job `lint`). Le §B de
  l'ADR-0008 reposait sur la vigilance humaine — c'est ainsi que ce bug a
  survecu des mois. Il est desormais outille.
- Tests 4 categories en egalite `Decimal` exacte. Les steps BDD toleraient
  jusqu'ici ±5 points de pourcentage sur un seuil fixe a 50%.

Verifie : `cargo check --tests` propre, 1665 tests unitaires verts,
clippy `-D warnings` exit 0, `bdd_governance` exit 0 (AG Visio 20/20),
`astro check` 0 erreur, vitest 352 verts, gate teste en positif ET negatif.

Refs: #661

+++ backend/src/application/dto/budget_dto.rs
@@ -82,30 +82,30 @@ pub struct BudgetVarianceResponse {
     pub building_id: Uuid,
 
     // Budget prévu
-    pub budgeted_ordinary: f64,
-    pub budgeted_extraordinary: f64,
-    pub budgeted_total: f64,
+    pub budgeted_ordinary: Decimal,
+    pub budgeted_extraordinary: Decimal,
+    pub budgeted_total: Decimal,
 
     // Réalisé (YTD - Year To Date)
-    pub actual_ordinary: f64,
-    pub actual_extraordinary: f64,
-    pub actual_total: f64,
+    pub actual_ordinary: Decimal,
+    pub actual_extraordinary: Decimal,
+    pub actual_total: Decimal,
 
     // Variance (budget - actual)
-    pub variance_ordinary: f64,
-    pub variance_extraordinary: f64,
-    pub variance_total: f64,
+    pub variance_ordinary: Decimal,
+    pub variance_extraordinary: Decimal,
+    pub variance_total: Decimal,
 
     // Variance en % ((budget - actual) / budget * 100)
-    pub variance_ordinary_pct: f64,
-    pub variance_extraordinary_pct: f64,
-    pub variance_total_pct: f64,
+    pub variance_ordinary_pct: Decimal,
+    pub variance_extraordinary_pct: Decimal,
+    pub variance_total_pct: Decimal,
 
     // Alertes
-    pub has_overruns: bool,              // Dépassements > 10%
-    pub overrun_categories: Vec<String>, // Catégories en dépassement
-    pub months_elapsed: i32,             // Mois écoulés dans l'exercice
-    pub projected_year_end_total: f64,   // Projection fin d'année
+    pub has_overruns: bool,                // Dépassements > 10%
+    pub overrun_categories: Vec<String>,   // Catégories en dépassement
+    pub months_elapsed: i32,               // Mois écoulés dans l'exercice
+    pub projected_year_end_total: Decimal, // Projection fin d'année
 }
 
 /// Response avec statistiques budgétaires
@@ -117,6 +117,7 @@ pub struct BudgetStatsResponse {
     pub approved_count: i64,
     pub rejected_count: i64,
     pub archived_count: i64,
-    pub average_total_budget: f64,
-    pub average_monthly_provision: f64,
+    // #661 — moyennes de montants : Decimal comme le reste du PCMN.
+    pub average_total_budget: Decimal,
+    pub average_monthly_provision: Decimal,
 }
